Rewrote the way graph functions are drawn
Now the functions will be saved on the graph instead of the DrawFunction method saving just the points that were drawn when the function was called, which meant that moving the camera around would not update the function drawn points.

Camera panning is also now possible with the Graph.Move function, with infinitely rendered function lines.